+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Fixes
 - **todo-commithash** Don't show update available prompt if the found version is the same as the running app version
+- [#120](https://github.com/dweymouth/supersonic/issues/120),[#87](https://github.com/dweymouth/supersonic/issues/87) Update Mac build process to support OS versions back to High Sierra (thanks @whorfin!)
 
 ## [0.1.0-beta] - 2023-04-01